+++ tests/src/test_tools.c
@@ -34,7 +34,7 @@ bool _expect_tokens(const char* source, const expect_token_t* expects, size_t ex
                     break;
                 }
                 case RLSL_TOKEN_LITERAL_FLOAT: {
-                    const long double epsilon = 0.000001L;
+                    const double epsilon = 0.000001;
                     if(fabsl(expect.value.d - token.value.num_float.value) > epsilon) {
                         goto fail;
                     }
